Move nativeEndian and isLinuxAWSUbuntu to common_linux
Only used by code in linux_bpf tag
