+++ pkg/ebpf/common.go
@@ -2,15 +2,12 @@ package ebpf
 
 import (
 	"bufio"
-	"encoding/binary"
 	"fmt"
-	"os"
-	"strings"
-	"unsafe"
-
 	"github.com/DataDog/datadog-agent/pkg/process/util"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 	"github.com/pkg/errors"
+	"os"
+	"strings"
 )
 
 // Feature versions sourced from: https://github.com/iovisor/bcc/blob/master/docs/kernel-versions.md
@@ -28,8 +25,6 @@ var requiredKernelFuncs = []string{
 var (
 	// ErrNotImplemented will be returned on non-linux environments like Windows and Mac OSX
 	ErrNotImplemented = errors.New("BPF-based system probe not implemented on non-linux systems")
-
-	nativeEndian binary.ByteOrder
 )
 
 func kernelCodeToString(code uint32) string {
@@ -107,25 +102,8 @@ func verifyKernelFuncs(path string) ([]string, error) {
 	return missing, nil
 }
 
-// In lack of binary.NativeEndian ...
-func init() {
-	var i int32 = 0x01020304
-	u := unsafe.Pointer(&i)
-	pb := (*byte)(u)
-	b := *pb
-	if b == 0x04 {
-		nativeEndian = binary.LittleEndian
-	} else {
-		nativeEndian = binary.BigEndian
-	}
-}
-
 func isUbuntu(platform string) bool {
 	return strings.Contains(strings.ToLower(platform), "ubuntu")
-}
-
-func isLinuxAWSUbuntu(platform string) bool {
-	return strings.Contains(strings.ToLower(platform), "aws") && isUbuntu(platform)
 }
 
 func isCentOS(platform string) bool {
