+++ pkg/ebpf/common_linux.go
@@ -1,0 +1,30 @@
+// +build linux_bpf
+
+package ebpf
+
+import (
+	"encoding/binary"
+	"strings"
+	"unsafe"
+)
+
+var (
+	nativeEndian binary.ByteOrder
+)
+
+// In lack of binary.NativeEndian ...
+func init() {
+	var i int32 = 0x01020304
+	u := unsafe.Pointer(&i)
+	pb := (*byte)(u)
+	b := *pb
+	if b == 0x04 {
+		nativeEndian = binary.LittleEndian
+	} else {
+		nativeEndian = binary.BigEndian
+	}
+}
+
+func isLinuxAWSUbuntu(platform string) bool {
+	return strings.Contains(strings.ToLower(platform), "aws") && isUbuntu(platform)
+}
